// src/app/models/pembayaran.ts
export interface pembayaran {
    tapper: any;

    checked: any;
    nama: string;
    smallHolderId: number;
    noPatG: string;
    noic: string;
    type: string;
    transactionId: number;
    dealerId: string;
    transDate: Date;

    weight: number;
    unitPrice: number;
    drc: number;
    total: number;
    noResit: string;
    companyAddress: string;

    paymentStatusId: number;
    dealerPekebunId: string;
    receiptNumber: string;

    printed: number;
    paymentDate: string;


}